prevent admin self-delete,-block or -admin-mode

+++ backend/backend/api/user.js
@@ -1315,6 +1315,14 @@ router.get('/users/:userId', getUser, isUserMasterAdmin, async function(
 router.delete('/:userId', getUser, isUserMasterAdmin, async function(req, res) {
     try {
         const userId = req.params.userId;
+        const authUserId = req.user.id;
+        if (userId === authUserId) {
+            const err = new Error(
+                "Invalid operation! You can't perform this operation on your own account"
+            );
+            err.code = 400;
+            throw err;
+        }
         const masterUserId = req.user.id || null;
         const user = await UserService.deleteBy({ _id: userId }, masterUserId);
         return sendItemResponse(req, res, user);
@@ -1329,6 +1337,14 @@ router.put('/:userId/restoreUser', getUser, isUserMasterAdmin, async function(
 ) {
     try {
         const userId = req.params.userId;
+        const authUserId = req.user.id;
+        if (userId === authUserId) {
+            const err = new Error(
+                "Invalid operation! You can't perform this operation on your own account"
+            );
+            err.code = 400;
+            throw err;
+        }
         const user = await UserService.restoreBy({
             _id: userId,
             deleted: true,
@@ -1345,6 +1361,14 @@ router.put('/:userId/blockUser', getUser, isUserMasterAdmin, async function(
 ) {
     try {
         const userId = req.params.userId;
+        const authUserId = req.user.id;
+        if (userId === authUserId) {
+            const err = new Error(
+                "Invalid operation! You can't perform this operation on your own account"
+            );
+            err.code = 400;
+            throw err;
+        }
         const user = await UserService.updateOneBy(
             { _id: userId },
             { isBlocked: true }
@@ -1361,6 +1385,14 @@ router.put('/:userId/unblockUser', getUser, isUserMasterAdmin, async function(
 ) {
     try {
         const userId = req.params.userId;
+        const authUserId = req.user.id;
+        if (userId === authUserId) {
+            const err = new Error(
+                "Invalid operation! You can't perform this operation on your own account"
+            );
+            err.code = 400;
+            throw err;
+        }
         const user = await UserService.updateOneBy(
             { _id: userId },
             { isBlocked: false }
@@ -1383,6 +1415,14 @@ router.post(
     async function(req, res) {
         try {
             const userId = req.params.userId;
+            const authUserId = req.user.id;
+            if (userId === authUserId) {
+                const err = new Error(
+                    "Invalid operation! You can't perform this operation on your own account"
+                );
+                err.code = 400;
+                throw err;
+            }
             const { temporaryPassword } = req.body;
             const user = await UserService.switchToAdminMode(
                 userId,
@@ -1407,6 +1447,14 @@ router.post(
     async function(req, res) {
         try {
             const userId = req.params.userId;
+            const authUserId = req.user.id;
+            if (userId === authUserId) {
+                const err = new Error(
+                    "Invalid operation! You can't perform this operation on your own account"
+                );
+                err.code = 400;
+                throw err;
+            }
             const user = await UserService.exitAdminMode(userId);
             return sendItemResponse(req, res, user);
         } catch (error) {
